Renamed frontend >> toasting
