+++ src/components/Header.tsx
@@ -7,7 +7,7 @@ import Link from "next/link";
 import clsx from "clsx";
 import Button from "@mui/material/Button";
 import ThemeSwitch from "../fragments/ThemeSwitch.Header";
-import { getAccountDisplayValue } from "../utils/frontend";
+import { getAccountDisplayValue } from "../utils/toasting";
 import { useAccount } from "../pages/_app";
 import { connected, connectWallet, network } from "../utils/metamask";
